requirements.txt
python-dotenv
sounddevice
numpy
scipy
openai
keyboard